refactor: seperate validateAndGetData and handleApiError to re-use
