+++ ui/src/context/review/review-context.tsx
@@ -1,5 +1,8 @@
 'use client';
 
+import { useToast } from '@/context/common/toast-context';
+import { useEditorRef } from '@/context/editor/editor-ref-context';
+import { validateReviewFields } from '@/lib/utils/review/validate-review-fields';
 import React, {
   Dispatch,
   ReactNode,
@@ -19,6 +22,13 @@ interface ContextShape {
   movieNameRef: React.RefObject<HTMLTextAreaElement>;
   disabled: boolean;
   setDisabled: Dispatch<SetStateAction<boolean>>;
+  validateAndGetData: () =>
+    | {
+        title: string;
+        movieName: string;
+        content: string;
+      }
+    | undefined;
 }
 
 const ReviewContext = createContext<ContextShape | null>(null);
@@ -41,6 +51,40 @@ export function ReviewProvider({
 
   const [disabled, setDisabled] = useState(false);
 
+  const { editorRef } = useEditorRef() ?? {};
+
+  const { emitToast } = useToast();
+
+  const validateAndGetData = () => {
+    if (!editorRef?.current) {
+      return;
+    }
+
+    const editorState = editorRef.current.getEditorState();
+
+    const validatedFields = validateReviewFields({ title, movieName, editorState });
+
+    if (!validatedFields.success) {
+      // Order to check should be aligned with the order of elements in the DOM
+      // title -> movieName -> editor
+      if (validatedFields.errors.title) {
+        titleRef.current?.focus();
+        emitToast(validatedFields.errors.title, 'error');
+      } else if (validatedFields.errors.movieName) {
+        movieNameRef.current?.focus();
+        emitToast(validatedFields.errors.movieName, 'error');
+      } else {
+        editorRef?.current?.focus();
+        emitToast(validatedFields.errors.content!, 'error');
+      }
+
+      setDisabled(false);
+      return;
+    }
+
+    return validatedFields.data;
+  };
+
   return (
     <ReviewContext.Provider
       value={{
@@ -52,6 +96,7 @@ export function ReviewProvider({
         movieNameRef,
         disabled,
         setDisabled,
+        validateAndGetData,
       }}
     >
       {children}
